+++ codes/GameAPIServer_Template/Services/Interfaces/IItemService.cs
@@ -1,4 +1,5 @@
 ﻿using GameAPIServer.Models;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -6,9 +7,9 @@ namespace GameAPIServer.Servicies.Interfaces;
 
 public interface IItemService
 {
-    public Task<(ErrorCode, List<RewardData>)> ReceiveOneGacha(int uid, int gachaKey);
+    public Task<(ErrorCode, List<RewardData>)> ReceiveOneGacha(Int64 uid, int gachaKey);
 
-    public Task<ErrorCode> ReceiveReward(int uid, RewardData reward);
+    public Task<ErrorCode> ReceiveReward(Int64 uid, RewardData reward);
   
 
 }
